chore: move `GardenIngressController` to standalone file
To avoid circular imports.
