+++ core/src/plugins/kubernetes/nginx/ingress-controller.ts
@@ -14,19 +14,7 @@ import { Microk8sGardenIngressController } from "./nginx-microk8s.js"
 import { MinikubeGardenIngressController } from "./nginx-minikube.js"
 import { KindGardenIngressController } from "./nginx-kind.js"
 import { EphemeralHelmGardenIngressController } from "./nginx-helm-ephemeral.js"
-import type { DeployState } from "../../../types/service.js"
-
-export abstract class GardenIngressController {
-  abstract install(ctx: KubernetesPluginContext, log: Log): Promise<void>
-
-  abstract uninstall(ctx: KubernetesPluginContext, log: Log): Promise<void>
-
-  async ready(ctx: KubernetesPluginContext, log: Log): Promise<boolean> {
-    return (await this.getStatus(ctx, log)) === "ready"
-  }
-
-  abstract getStatus(ctx: KubernetesPluginContext, log: Log): Promise<DeployState>
-}
+import type { GardenIngressController } from "./ingress-controller-base.js"
 
 export function getGardenIngressController(ctx: KubernetesPluginContext): GardenIngressController | undefined {
   const clusterType = ctx.provider.config.clusterType
